+++ UI/ChallengeDetails/Tabs/Summary.lua
@@ -37,13 +37,8 @@ function MyDungeonsBook:SummaryFrame_Create(parentFrame, challengeId)
 			end
 		end
 	end
-	local deathsGraph, scaleSlider, timelineSlider = self:SingleIconsGraph_Create(summaryFrame, "DeathsGraph", series, challenge.challengeInfo.startTime, xLimit, 150, legend);
-	scaleSlider.slider:Hide();
-	scaleSlider.label:Hide();
-	timelineSlider.slider:Hide();
+	local deathsGraph = self:SingleIconsGraph_Create(summaryFrame, "DeathsGraph", series, challenge.challengeInfo.startTime, xLimit, 150, legend);
 	summaryFrame:SetUserData("deathsGraph", deathsGraph);
-	summaryFrame:SetUserData("scaleSlider", scaleSlider);
-	summaryFrame:SetUserData("timelineSlider", timelineSlider);
 	local table = self:TableWidget_Create(cols, 5, 40, nil, summaryFrame, "summary");
 	summaryFrame:SetUserData("summaryTable", table);
 	table:SetData(self:SummaryFrame_GetDataForTable(challengeId));
@@ -76,13 +71,10 @@ function MyDungeonsBook:SummaryFrame_Create(parentFrame, challengeId)
 		end
 	});
 	table:SortData();
-	table.frame:SetPoint("TOPLEFT", 0, -350);
+	table.frame:SetPoint("TOPLEFT", 0, -300);
 	summaryFrame:SetCallback("OnRelease", function(frame)
 		frame:GetUserData("deathsGraph"):ResetData();
 		frame:GetUserData("deathsGraph"):Hide();
-		frame:GetUserData("scaleSlider").slider:Show();
-		frame:GetUserData("scaleSlider").label:Show();
-		frame:GetUserData("timelineSlider").slider:Show();
 		frame:GetUserData("summaryTable"):Hide();
 	end);
 	return summaryFrame;